refactor: ai_response_generator.py を ai.py に、message_repository.py を db.py にリネームし、参照を更新
